pyproject.toml
[tool.poetry]
name = "photomanager"
version = "0.1.0"
description = ""
authors = ["arturkalimullin <noti1996@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
psycopg2-binary = "^2.9.5"
python-dotenv = "^0.21.0"
djangorestframework = "^3.14.0"
drf-yasg = "^1.21.4"
pillow = "^9.3.0"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
isort = "^5.11.2"
ipython = "^8.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
